+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
-- Children in `storeWrapper` to consume PWA Context. 
+### Removed
+- PWA helmet and query in `StoreWrapper` and move it to `PWAProvider`. 
 
 ## [2.52.3] - 2019-08-27
 ### Fixed